Move everything but RMA from MPI-CUDA to a new host-transfer backend.
