+++ benchmark/benchmark_allreduces.cpp
@@ -31,8 +31,8 @@
 #ifdef AL_HAS_NCCL
 #include "test_utils_nccl_cuda.hpp"
 #endif
-#ifdef AL_HAS_MPI_CUDA
-#include "test_utils_mpi_cuda.hpp"
+#ifdef AL_HAS_HOST_TRANSFER
+#include "test_utils_ht.hpp"
 #endif
 
 size_t start_size = 1;
@@ -118,8 +118,13 @@ int main(int argc, char** argv) {
 #endif    
 #ifdef AL_HAS_MPI_CUDA
   } else if (backend == "MPI-CUDA") {
-    do_benchmark<Al::MPICUDABackend>(sizes);
+    std::cout << "Allreduce not supported on MPI-CUDA backend." << std::endl;
+    std::abort();
 #endif    
+#ifdef AL_HAS_HOST_TRANSFER
+  } else if (backend == "HT") {
+    do_benchmark<Al::HostTransferBackend>(sizes);
+#endif
   } else {
     std::cerr << "usage: " << argv[0] << " [MPI";
 #ifdef AL_HAS_NCCL
@@ -127,6 +132,9 @@ int main(int argc, char** argv) {
 #endif
 #ifdef AL_HAS_MPI_CUDA
     std::cerr << " | MPI-CUDA";
+#endif
+#ifdef AL_HAS_HOST_TRANSFER
+    std::cerr << " | HT";
 #endif
     std::cerr << "]" << std::endl;
     return -1;
