+++ benchmark/benchmark_nballreduces.cpp
@@ -31,8 +31,8 @@
 #ifdef AL_HAS_NCCL
 #include "test_utils_nccl_cuda.hpp"
 #endif
-#ifdef AL_HAS_MPI_CUDA
-#include "test_utils_mpi_cuda.hpp"
+#ifdef AL_HAS_HOST_TRANSFER
+#include "test_utils_ht.hpp"
 #endif
 
 size_t start_size = 1;
@@ -59,10 +59,10 @@ typename Al::NCCLBackend::comm_type get_comm<Al::NCCLBackend>() {
 }
 #endif
 
-#ifdef AL_HAS_MPI_CUDA
+#ifdef AL_HAS_HOST_TRANSFER
 template <>
-typename Al::MPICUDABackend::comm_type get_comm<Al::MPICUDABackend>() {
-  return typename Al::MPICUDABackend::comm_type(MPI_COMM_WORLD, bm_stream);
+typename Al::HostTransferBackend::comm_type get_comm<Al::HostTransferBackend>() {
+  return typename Al::HostTransferBackend::comm_type(MPI_COMM_WORLD, bm_stream);
 }
 #endif
 
@@ -258,8 +258,13 @@ int main(int argc, char** argv) {
 #endif    
 #ifdef AL_HAS_MPI_CUDA
   } else if (backend == "MPI-CUDA") {
-    do_benchmark<Al::MPICUDABackend>();
+    std::cout << "Allreduce not supported on MPI-CUDA backend." << std::endl;
+    std::abort();
 #endif    
+#ifdef AL_HAS_HOST_TRANSFER
+  } else if (backend == "HT") {
+    do_benchmark<Al::HostTransferBackend>();
+#endif
   } else {
     std::cerr << "usage: " << argv[0] << " [MPI";
 #ifdef AL_HAS_NCCL
@@ -267,6 +272,9 @@ int main(int argc, char** argv) {
 #endif
 #ifdef AL_HAS_MPI_CUDA
     std::cerr << " | MPI-CUDA";
+#endif
+#ifdef AL_HAS_HOST_TRANSFER
+    std::cerr << " | HT";
 #endif
     std::cerr << "]" << std::endl;
     return -1;
