+++ cmake/Al_config.hpp.in
@@ -2,6 +2,7 @@
 
 #cmakedefine AL_HAS_CUDA
 #cmakedefine AL_HAS_MPI_CUDA
+#cmakedefine AL_HAS_HOST_TRANSFER
 #cmakedefine AL_HAS_NCCL
 #cmakedefine AL_HAS_ROCM
