+++ cmake/AluminumConfig.cmake.in
@@ -8,6 +8,7 @@ set(ALUMINUM_VERSION ${PACKAGE_VERSION})
 set(AL_HAS_CUDA @AL_HAS_CUDA@)
 set(AL_HAS_ROCM @AL_HAS_ROCM@)
 set(AL_HAS_MPI_CUDA @AL_HAS_MPI_CUDA@)
+set(AL_HAS_HOST_TRANSFER @AL_HAS_HOST_TRANSFER@)
 set(AL_HAS_NCCL @AL_HAS_NCCL@)
 
 set(MPI_CXX_COMPILER "@MPI_CXX_COMPILER@" CACHE FILEPATH
