+++ include/aluminum/ht/reduce_scatter.hpp
@@ -28,49 +28,42 @@
 #pragma once
 
 #include "cuda.hpp"
-#include "mpi_cuda/communicator.hpp"
+#include "ht/communicator.hpp"
 #include "progress.hpp"
 
 namespace Al {
 namespace internal {
-namespace mpi_cuda {
+namespace ht {
 
 template <typename T>
-class GatherAlState : public AlState {
+class ReduceScatterAlState : public AlState {
 public:
-  GatherAlState(const T* sendbuf, T* recvbuf, size_t count, int root,
-                MPICUDACommunicator& comm, cudaStream_t stream) :
+  ReduceScatterAlState(const T* sendbuf, T* recvbuf, size_t count,
+                       ReductionOperator op, HostTransferCommunicator& comm,
+                       cudaStream_t stream) :
     AlState(nullptr),
-    rank_(comm.rank()), root_(root), count_(count),
-    host_mem_(get_pinned_memory<T>(rank_ == root_ ? comm.size()*count_ : count_)),
+    count_(count),
+    host_mem_(get_pinned_memory<T>(comm.size()*count_)),
+    op_(mpi::ReductionOperator2MPI_Op(op)),
     comm_(comm.get_comm()),
     compute_stream(comm.get_stream()) {
 
-    bool const i_am_root = rank_ == root_;
-    bool const inplace_operation = sendbuf == recvbuf;
-
     // Transfer data from device to host and use an event to determine when it
     // completes.
-    if (i_am_root)
-      AL_CHECK_CUDA(cudaMemcpyAsync(
-                      host_mem_+rank_*count_,
-                      inplace_operation ? sendbuf+rank_*count_ : sendbuf,
-                      sizeof(T)*count, cudaMemcpyDeviceToHost, stream));
-    else
-      AL_CHECK_CUDA(cudaMemcpyAsync(host_mem_, sendbuf, sizeof(T)*count_,
-                                    cudaMemcpyDeviceToHost, stream));
+    AL_CHECK_CUDA(cudaMemcpyAsync(host_mem_, sendbuf, sizeof(T)*count*comm.size(),
+                                  cudaMemcpyDeviceToHost, stream));
     d2h_event_.record(stream);
+
+    // Have the device wait on the host.
     gpuwait_.wait(stream);
 
-    if (i_am_root) {
-      // Transfer completed buffer back to device.
-      AL_CHECK_CUDA(cudaMemcpyAsync(recvbuf, host_mem_, sizeof(T)*count*comm.size(),
-                                    cudaMemcpyHostToDevice, stream));
-      h2d_event_.record(stream);
-    }
+    // Transfer completed buffer back to device.
+    AL_CHECK_CUDA(cudaMemcpyAsync(recvbuf, host_mem_, sizeof(T)*count,
+                                  cudaMemcpyHostToDevice, stream));
+    h2d_event_.record(stream);
   }
 
-  ~GatherAlState() override {
+  ~ReduceScatterAlState() override {
     release_pinned_memory(host_mem_);
   }
 
@@ -83,74 +76,47 @@ public:
         return PEAction::cont;
       }
     }
-    if (!gather_started_) {
-      if (root_ == rank_) {
-        MPI_Igather(MPI_IN_PLACE, count_, mpi::TypeMap<T>(),
-                    host_mem_, count_, mpi::TypeMap<T>(),
-                    root_, comm_, &req_);
-      } else {
-        MPI_Igather(host_mem_, count_, mpi::TypeMap<T>(),
-                    host_mem_, count_, mpi::TypeMap<T>(),
-                    root_, comm_, &req_);
-        gpuwait_.signal();
-      }
-      gather_started_ = true;
+    if (!rs_started_) {
+      MPI_Ireduce_scatter_block(MPI_IN_PLACE, host_mem_, count_,
+                                mpi::TypeMap<T>(), op_, comm_, &req_);
+      rs_started_ = true;
     }
-
-    if (!gather_done_) {
-      // Wait for the gather to complete
+    if (!rs_done_) {
+      // Wait for the RS to complete.
       int flag;
       MPI_Test(&req_, &flag, MPI_STATUS_IGNORE);
       if (flag) {
-        gather_done_ = true;
-        if (rank_ == root_)
-          gpuwait_.signal();
-        else
-          return PEAction::complete;
-      }
-      else {
+        rs_done_ = true;
+        gpuwait_.signal();
+      } else {
         return PEAction::cont;
       }
     }
-    else if (rank_ != root_) {
-      // Paranoia, in case step() is ever called again after returning
-      // 'true' for the first time.
-      return PEAction::complete;
-    }
-
     // Wait for host-to-device memcopy; cleanup
     if (h2d_event_.query()) {
       return PEAction::complete;
     }
-
     return PEAction::cont;
   }
 
   bool needs_completion() const override { return false; }
   void* get_compute_stream() const override { return compute_stream; }
-
-  std::string get_name() const override { return "HTGather"; }
+  std::string get_name() const override { return "HTReduceScatter"; }
 
 private:
-  int rank_;
-  int root_;
   size_t count_;
   T* host_mem_;
-
   cuda::GPUWait gpuwait_;
-
   cuda::FastEvent d2h_event_, h2d_event_;
-
+  MPI_Op op_;
   MPI_Comm comm_;
   MPI_Request req_ = MPI_REQUEST_NULL;
-
   bool mem_xfer_done_ = false;
-  bool gather_started_ = false;
-  bool gather_done_ = false;
-
+  bool rs_started_ = false;
+  bool rs_done_ = false;
   cudaStream_t compute_stream;
 };
 
-}  // namespace mpi_cuda
+}  // namespace ht
 }  // namespace internal
 }  // namespace Al
