+++ include/aluminum/mpi_cuda_impl.hpp
@@ -28,50 +28,12 @@
 #pragma once
 
 #include "Al.hpp"
-#include "mpi_cuda/allgather.hpp"
-#include "mpi_cuda/allreduce.hpp"
-#include "mpi_cuda/alltoall.hpp"
-#include "mpi_cuda/bcast.hpp"
-#include "mpi_cuda/gather.hpp"
-#include "mpi_cuda/reduce.hpp"
-#include "mpi_cuda/reduce_scatter.hpp"
-#include "mpi_cuda/scatter.hpp"
-
-#include "mpi_cuda/pt2pt.hpp"
+#include "mpi_cuda/communicator.hpp"
 #ifdef AL_HAS_MPI_CUDA_RMA
 #include "mpi_cuda/rma.hpp"
 #endif
 
 namespace Al {
-
-enum class MPICUDAAllreduceAlgorithm {
-  automatic,
-  host_transfer
-};
-
-enum class MPICUDACollectiveAlgorithm {
-  automatic
-};
-
-inline std::string algorithm_name(MPICUDAAllreduceAlgorithm algo) {
-  switch (algo) {
-  case MPICUDAAllreduceAlgorithm::automatic:
-    return "automatic";
-  case MPICUDAAllreduceAlgorithm::host_transfer:
-    return "host-transfer";
-  default:
-    return "unknown";
-  }
-}
-
-inline std::string algorithm_name(MPICUDACollectiveAlgorithm algo) {
-  switch (algo) {
-  case MPICUDACollectiveAlgorithm::automatic:
-    return "automatic";
-  default:
-    return "unknown";
-  }
-}
 
 namespace internal {
 namespace mpi_cuda {
@@ -100,117 +62,9 @@ class MPICUDABackend {
   friend void internal::mpi_cuda::init(int&, char**&);
   friend void internal::mpi_cuda::finalize();
  public:
-  using allreduce_algo_type = MPICUDAAllreduceAlgorithm;
-  using allgather_algo_type = MPICUDACollectiveAlgorithm;
-  using alltoall_algo_type = MPICUDACollectiveAlgorithm;
-  using bcast_algo_type = MPICUDACollectiveAlgorithm;
-  using gather_algo_type = MPICUDACollectiveAlgorithm;
-  using reduce_algo_type = MPICUDACollectiveAlgorithm;
-  using reduce_scatter_algo_type = MPICUDACollectiveAlgorithm;
-  using scatter_algo_type = MPICUDACollectiveAlgorithm;
   using comm_type = internal::mpi_cuda::MPICUDACommunicator;
   using req_type = std::shared_ptr<internal::mpi_cuda::MPICUDARequest>;
   static constexpr std::nullptr_t null_req = nullptr;
-
-  template <typename T>
-  static void Allreduce(const T* sendbuf, T* recvbuf, size_t count,
-                        ReductionOperator op, comm_type& comm,
-                        allreduce_algo_type algo) {
-    if (count == 0) return;
-    switch (algo) {
-    case MPICUDAAllreduceAlgorithm::automatic:
-    case MPICUDAAllreduceAlgorithm::host_transfer:
-      do_host_transfer_allreduce(sendbuf, recvbuf, count, op, comm,
-                                 comm.get_stream());
-      break;
-    default:
-      throw_al_exception("Invalid algorithm");
-    }
-  }
-  template <typename T>
-  static void Allreduce(T* recvbuf, size_t count,
-                        ReductionOperator op, comm_type& comm,
-                        allreduce_algo_type algo) {
-    Allreduce(recvbuf, recvbuf, count, op, comm, algo);
-  }
-
-  template <typename T>
-  static void NonblockingAllreduce(
-      const T* sendbuf, T* recvbuf, size_t count,
-      ReductionOperator op,
-      comm_type& comm,
-      req_type& req,
-      allreduce_algo_type algo) {
-    if (count == 0) return;
-    cudaStream_t internal_stream = internal::cuda::get_internal_stream();
-    sync_internal_stream_with_comm(internal_stream, comm);
-    switch (algo) {
-    case MPICUDAAllreduceAlgorithm::automatic:
-    case MPICUDAAllreduceAlgorithm::host_transfer:
-      do_host_transfer_allreduce(sendbuf, recvbuf, count, op, comm,
-                                 internal_stream);
-      break;
-    default:
-      throw_al_exception("Invalid algorithm");
-    }
-    setup_completion_event(internal_stream, comm, req);
-  }
-
-  template <typename T>
-  static void NonblockingAllreduce(
-      T* recvbuf, size_t count,
-      ReductionOperator op, comm_type& comm,
-      req_type& req,
-      allreduce_algo_type algo) {
-    NonblockingAllreduce(recvbuf, recvbuf, count, op, comm,
-                         req, algo);
-  }
-
-  template <typename T>
-  static void Send(const T* sendbuf, size_t count, int dest, comm_type& comm) {
-    do_send(sendbuf, count, dest, comm, comm.get_stream());
-  }
-
-  template <typename T>
-  static void NonblockingSend(const T* sendbuf, size_t count, int dest,
-                              comm_type& comm, req_type& req) {
-    cudaStream_t internal_stream = internal::cuda::get_internal_stream();
-    sync_internal_stream_with_comm(internal_stream, comm);
-    do_send(sendbuf, count, dest, comm, internal_stream);
-    setup_completion_event(internal_stream, comm, req);
-  }
-
-  template <typename T>
-  static void Recv(T* recvbuf, size_t count, int src, comm_type& comm) {
-    do_recv(recvbuf, count, src, comm, comm.get_stream());
-  }
-
-  template <typename T>
-  static void NonblockingRecv(T* recvbuf, size_t count, int src,
-                              comm_type& comm, req_type& req) {
-    cudaStream_t internal_stream = internal::cuda::get_internal_stream();
-    sync_internal_stream_with_comm(internal_stream, comm);
-    do_recv(recvbuf, count, src, comm, internal_stream);
-    setup_completion_event(internal_stream, comm, req);
-  }
-
-  template <typename T>
-  static void SendRecv(const T* sendbuf, size_t send_count, int dest,
-                       T* recvbuf, size_t recv_count, int src, comm_type& comm) {
-    do_sendrecv(sendbuf, send_count, dest, recvbuf, recv_count, src, comm,
-                comm.get_stream());
-  }
-
-  template <typename T>
-  static void NonblockingSendRecv(const T* sendbuf, size_t send_count, int dest,
-                                  T* recvbuf, size_t recv_count, int src,
-                                  comm_type& comm, req_type& req) {
-    cudaStream_t internal_stream = internal::cuda::get_internal_stream();
-    sync_internal_stream_with_comm(internal_stream, comm);
-    do_sendrecv(sendbuf, send_count, dest, recvbuf, recv_count, src, comm,
-                internal_stream);
-    setup_completion_event(internal_stream, comm, req);
-  }
 
 #ifdef AL_HAS_MPI_CUDA_RMA
   template <typename T>
@@ -248,446 +102,7 @@ class MPICUDABackend {
   }
 #endif // AL_HAS_MPI_CUDA_RMA
 
-  template <typename T>
-  static void Allgather(
-    const T* sendbuf, T* recvbuf, size_t count,
-    comm_type& comm, allgather_algo_type algo) {
-    if (count == 0) return;
-    switch (algo) {
-    case MPICUDACollectiveAlgorithm::automatic:
-      do_allgather(sendbuf, recvbuf, count, comm, comm.get_stream());
-      break;
-    default:
-      throw_al_exception("MPICUDAAllgather: Invalid algorithm");
-    }
-  }
-
-  template <typename T>
-  static void Allgather(
-    T* buffer, size_t count, comm_type& comm, allgather_algo_type algo) {
-    Allgather(buffer, buffer, count, comm, algo);
-  }
-
-  template <typename T>
-  static void NonblockingAllgather(
-    const T* sendbuf, T* recvbuf, size_t count,
-    comm_type& comm, req_type& req, allgather_algo_type algo) {
-    if (count == 0) return;
-    cudaStream_t internal_stream = internal::cuda::get_internal_stream();
-    sync_internal_stream_with_comm(internal_stream, comm);
-    switch (algo) {
-    case MPICUDACollectiveAlgorithm::automatic:
-      do_allgather(sendbuf, recvbuf, count, comm, internal_stream);
-      break;
-    default:
-      throw_al_exception("MPICUDAAllgather: Invalid algorithm");
-    }
-    setup_completion_event(internal_stream, comm, req);
-  }
-
-  template <typename T>
-  static void NonblockingAllgather(
-    T* buffer, size_t count,
-    comm_type& comm, req_type& req, allgather_algo_type algo) {
-    NonblockingAllgather<T>(buffer, buffer, count, comm, req, algo);
-  }
-
-  template <typename T>
-  static void Alltoall(
-    const T* sendbuf, T* recvbuf, size_t count,
-    comm_type& comm, alltoall_algo_type algo) {
-    if (count == 0) return;
-    switch (algo) {
-    case MPICUDACollectiveAlgorithm::automatic:
-      do_alltoall(sendbuf, recvbuf, count, comm, comm.get_stream());
-      break;
-    default:
-      throw_al_exception("MPICUDAAlltoall: Invalid algorithm");
-    }
-  }
-
-  template <typename T>
-  static void Alltoall(
-    T* buffer, size_t count, comm_type& comm, alltoall_algo_type algo) {
-    Alltoall(buffer, buffer, count, comm, algo);
-  }
-
-  template <typename T>
-  static void NonblockingAlltoall(
-    const T* sendbuf, T* recvbuf, size_t count,
-    comm_type& comm, req_type& req, alltoall_algo_type algo) {
-    if (count == 0) return;
-    cudaStream_t internal_stream = internal::cuda::get_internal_stream();
-    sync_internal_stream_with_comm(internal_stream, comm);
-    switch (algo) {
-    case MPICUDACollectiveAlgorithm::automatic:
-      do_alltoall(sendbuf, recvbuf, count, comm, internal_stream);
-      break;
-    default:
-      throw_al_exception("Invalid algorithm");
-    }
-    setup_completion_event(internal_stream, comm, req);
-  }
-
-  template <typename T>
-  static void NonblockingAlltoall(
-    T* buffer, size_t count,
-    comm_type& comm, req_type& req, alltoall_algo_type algo) {
-    NonblockingAlltoall<T>(buffer, buffer, count, comm, req, algo);
-  }
-
-  template <typename T>
-  static void Bcast(T* buf, size_t count, int root, comm_type& comm,
-                    bcast_algo_type algo) {
-    if (count == 0) return;
-    switch (algo) {
-    case MPICUDACollectiveAlgorithm::automatic:
-      do_bcast(buf, count, root, comm, comm.get_stream());
-      break;
-    default:
-      throw_al_exception("MPICUDABcast: Invalid algorithm");
-    }
-  }
-
-  template <typename T>
-  static void NonblockingBcast(
-    T* buf, size_t count, int root,
-    comm_type& comm, req_type& req, bcast_algo_type algo) {
-    if (count == 0) return;
-    cudaStream_t internal_stream = internal::cuda::get_internal_stream();
-    sync_internal_stream_with_comm(internal_stream, comm);
-    switch (algo) {
-    case MPICUDACollectiveAlgorithm::automatic:
-      do_bcast(buf, count, root, comm, internal_stream);
-      break;
-    default:
-      throw_al_exception("MPICUDABcast: Invalid algorithm");
-    }
-    setup_completion_event(internal_stream, comm, req);
-  }
-
-  template <typename T>
-  static void Gather(
-    const T* sendbuf, T* recvbuf, size_t count, int root,
-    comm_type& comm, gather_algo_type algo) {
-    if (count == 0) return;
-    switch (algo) {
-    case MPICUDACollectiveAlgorithm::automatic:
-      do_gather(sendbuf, recvbuf, count, root, comm, comm.get_stream());
-      break;
-    default:
-      throw_al_exception("MPICUDAGather: Invalid algorithm");
-    }
-  }
-
-  template <typename T>
-  static void Gather(
-    T* buffer, size_t count, int root, comm_type& comm, gather_algo_type algo) {
-    Gather(buffer, buffer, count, root, comm, algo);
-  }
-
-  template <typename T>
-  static void NonblockingGather(
-    const T* sendbuf, T* recvbuf, size_t count, int root,
-    comm_type& comm, req_type& req, gather_algo_type algo) {
-    if (count == 0) return;
-    cudaStream_t internal_stream = internal::cuda::get_internal_stream();
-    sync_internal_stream_with_comm(internal_stream, comm);
-    switch (algo) {
-    case MPICUDACollectiveAlgorithm::automatic:
-      do_gather(sendbuf, recvbuf, count, root, comm, internal_stream);
-      break;
-    default:
-      throw_al_exception("Invalid algorithm");
-    }
-    setup_completion_event(internal_stream, comm, req);
-  }
-
-  template <typename T>
-  static void NonblockingGather(
-    T* buffer, size_t count, int root,
-    comm_type& comm, req_type& req, gather_algo_type algo) {
-    NonblockingGather<T>(buffer, buffer, count, root, comm, req, algo);
-  }
-
-  template <typename T>
-  static void Reduce(
-    const T* sendbuf, T* recvbuf, size_t count, ReductionOperator op, int root,
-    comm_type& comm, reduce_algo_type algo) {
-    if (count == 0) return;
-    switch (algo) {
-    case MPICUDACollectiveAlgorithm::automatic:
-      do_reduce(sendbuf, recvbuf, count, op, root, comm, comm.get_stream());
-      break;
-    default:
-      throw_al_exception("MPICUDAReduce: Invalid algorithm");
-    }
-  }
-
-  template <typename T>
-  static void Reduce(
-    T* buffer, size_t count, ReductionOperator op, int root, comm_type& comm,
-    reduce_algo_type algo) {
-    Reduce(buffer, buffer, count, op, root, comm, algo);
-  }
-
-  template <typename T>
-  static void NonblockingReduce(
-    const T* sendbuf, T* recvbuf, size_t count, ReductionOperator op, int root,
-    comm_type& comm, req_type& req, reduce_algo_type algo) {
-    if (count == 0) return;
-    cudaStream_t internal_stream = internal::cuda::get_internal_stream();
-    sync_internal_stream_with_comm(internal_stream, comm);
-    switch (algo) {
-    case MPICUDACollectiveAlgorithm::automatic:
-      do_reduce(sendbuf, recvbuf, count, op, root, comm, internal_stream);
-      break;
-    default:
-      throw_al_exception("Invalid algorithm");
-    }
-    setup_completion_event(internal_stream, comm, req);
-  }
-
-  template <typename T>
-  static void NonblockingReduce(
-    T* buffer, size_t count, ReductionOperator op, int root,
-    comm_type& comm, req_type& req, reduce_algo_type algo) {
-    NonblockingReduce<T>(buffer, buffer, count, op, root, comm, req, algo);
-  }
-
-  template <typename T>
-  static void Reduce_scatter(
-    const T* sendbuf, T* recvbuf, size_t count, ReductionOperator op,
-    comm_type& comm, reduce_scatter_algo_type algo) {
-    if (count == 0) return;
-    switch (algo) {
-    case MPICUDACollectiveAlgorithm::automatic:
-      do_reduce_scatter(sendbuf, recvbuf, count, op, comm, comm.get_stream());
-      break;
-    default:
-      throw_al_exception("MPICUDAReduce_scatter: Invalid algorithm");
-    }
-  }
-
-  template <typename T>
-  static void Reduce_scatter(
-    T* buffer, size_t count, ReductionOperator op,
-    comm_type& comm, reduce_scatter_algo_type algo) {
-    Reduce_scatter(buffer, buffer, count, op, comm, algo);
-  }
-
-  template <typename T>
-  static void NonblockingReduce_scatter(
-    const T* sendbuf, T* recvbuf, size_t count, ReductionOperator op,
-    comm_type& comm, req_type& req, reduce_scatter_algo_type algo) {
-    if (count == 0) return;
-    cudaStream_t internal_stream = internal::cuda::get_internal_stream();
-    sync_internal_stream_with_comm(internal_stream, comm);
-    switch (algo) {
-    case MPICUDACollectiveAlgorithm::automatic:
-      do_reduce_scatter(sendbuf, recvbuf, count, op, comm, internal_stream);
-      break;
-    default:
-      throw_al_exception("MPICUDAReduce_scatter: Invalid algorithm");
-    }
-    setup_completion_event(internal_stream, comm, req);
-  }
-
-  template <typename T>
-  static void NonblockingReduce_scatter(
-    T* buffer, size_t count, ReductionOperator op,
-    comm_type& comm, req_type& req, reduce_scatter_algo_type algo) {
-    NonblockingReduce_scatter(buffer, buffer, count, op, comm, req, algo);
-  }
-
-  template <typename T>
-  static void Scatter(
-    const T* sendbuf, T* recvbuf, size_t count, int root,
-    comm_type& comm, scatter_algo_type algo) {
-    if (count == 0) return;
-    switch (algo) {
-    case MPICUDACollectiveAlgorithm::automatic:
-      do_scatter(sendbuf, recvbuf, count, root, comm, comm.get_stream());
-      break;
-    default:
-      throw_al_exception("MPICUDAScatter: Invalid algorithm");
-    }
-  }
-
-  template <typename T>
-  static void Scatter(
-    T* buffer, size_t count, int root, comm_type& comm, scatter_algo_type algo) {
-    Scatter(buffer, buffer, count, root, comm, algo);
-  }
-
-  template <typename T>
-  static void NonblockingScatter(
-    const T* sendbuf, T* recvbuf, size_t count, int root,
-    comm_type& comm, req_type& req, scatter_algo_type algo) {
-    if (count == 0) return;
-    cudaStream_t internal_stream = internal::cuda::get_internal_stream();
-    sync_internal_stream_with_comm(internal_stream, comm);
-    switch (algo) {
-    case MPICUDACollectiveAlgorithm::automatic:
-      do_scatter(sendbuf, recvbuf, count, root, comm, internal_stream);
-      break;
-    default:
-      throw_al_exception("Invalid algorithm");
-    }
-    setup_completion_event(internal_stream, comm, req);
-  }
-
-  template <typename T>
-  static void NonblockingScatter(
-    T* buffer, size_t count, int root,
-    comm_type& comm, req_type& req, scatter_algo_type algo) {
-    NonblockingScatter<T>(buffer, buffer, count, root, comm, req, algo);
-  }
-
   static std::string Name() { return "MPICUDABackend"; }
-
- private:
-  /** Event for synchronizing between streams. */
-  static cudaEvent_t sync_event;
-  /**
-   * Set up stream synchronization.
-   * This will cause the provided internal stream to synchronize with the stream
-   * associated with comm.
-   */
-  static void sync_internal_stream_with_comm(
-    cudaStream_t internal_stream, comm_type& comm) {
-    // We can reuse a single event for cudaStreamWaitEvent because it uses the
-    // stream/event state when it is called.
-    AL_CHECK_CUDA(cudaEventRecord(sync_event, comm.get_stream()));
-    AL_CHECK_CUDA(cudaStreamWaitEvent(internal_stream, sync_event, 0));
-  }
-  /**
-   * Set up the request for completion checking.
-   * This uses an event recorded on the provided internal stream.
-   */
-  static void setup_completion_event(
-    cudaStream_t internal_stream, comm_type& comm, req_type& req) {
-    cudaEvent_t event = internal::cuda::get_cuda_event();
-    AL_CHECK_CUDA(cudaEventRecord(event, internal_stream));
-    req = std::make_shared<internal::mpi_cuda::MPICUDARequest>(event, comm.get_stream());
-  }
-
-  /** Run a host-transfer allreduce. */
-  template <typename T>
-  static void do_host_transfer_allreduce(
-    const T* sendbuf, T* recvbuf, size_t count, ReductionOperator op,
-    comm_type& comm, cudaStream_t internal_stream) {
-#if 0
-    if (!internal::cuda::stream_memory_operations_supported()) {
-      throw_al_exception("Host-transfer allreduce not supported without stream memory operations");
-    }
-#endif // 0
-    internal::mpi_cuda::HostTransferState<T>* state = new internal::mpi_cuda::HostTransferState<T>(
-      sendbuf, recvbuf, count, op, comm, internal_stream, internal::get_free_request());
-    internal::get_progress_engine()->enqueue(state);
-  }
-
-  template <typename T>
-  static void do_send(
-    const T* sendbuf, size_t count, int dest, comm_type& comm,
-    cudaStream_t stream) {
-    internal::mpi_cuda::SendAlState<T>* state =
-      new internal::mpi_cuda::SendAlState<T>(
-        sendbuf, count, dest, comm, stream);
-    internal::get_progress_engine()->enqueue(state);
-  }
-
-  template <typename T>
-  static void do_recv(
-    T* recvbuf, size_t count, int src, comm_type& comm,
-    cudaStream_t stream) {
-    internal::mpi_cuda::RecvAlState<T>* state =
-      new internal::mpi_cuda::RecvAlState<T>(
-        recvbuf, count, src, comm, stream);
-    internal::get_progress_engine()->enqueue(state);
-  }
-
-  template <typename T>
-  static void do_sendrecv(
-    const T* sendbuf, size_t send_count, int dest,
-    T* recvbuf, size_t recv_count, int src, comm_type& comm,
-    cudaStream_t stream) {
-    internal::mpi_cuda::SendRecvAlState<T>* state =
-      new internal::mpi_cuda::SendRecvAlState<T>(
-        sendbuf, send_count, dest, recvbuf, recv_count, src, comm, stream);
-    internal::get_progress_engine()->enqueue(state);
-  }
-
-  template <typename T>
-  static void do_allgather(
-    const T* sendbuf, T* recvbuf, size_t count, comm_type& comm,
-    cudaStream_t stream) {
-    internal::mpi_cuda::AllgatherAlState<T>* state =
-      new internal::mpi_cuda::AllgatherAlState<T>(
-        sendbuf, recvbuf, count, comm, stream);
-    internal::get_progress_engine()->enqueue(state);
-  }
-
-  template <typename T>
-  static void do_alltoall(
-    const T* sendbuf, T* recvbuf, size_t count, comm_type& comm,
-    cudaStream_t stream) {
-    internal::mpi_cuda::AlltoallAlState<T>* state =
-      new internal::mpi_cuda::AlltoallAlState<T>(
-        sendbuf, recvbuf, count, comm, stream);
-    internal::get_progress_engine()->enqueue(state);
-  }
-
-  template <typename T>
-  static void do_bcast(
-    T* buf, size_t count, int root, comm_type& comm,
-    cudaStream_t stream) {
-    internal::mpi_cuda::BcastAlState<T>* state =
-      new internal::mpi_cuda::BcastAlState<T>(
-        buf, count, root, comm, stream);
-    internal::get_progress_engine()->enqueue(state);
-  }
-
-  template <typename T>
-  static void do_gather(
-    const T* sendbuf, T* recvbuf, size_t count, int root, comm_type& comm,
-    cudaStream_t stream) {
-    internal::mpi_cuda::GatherAlState<T>* state =
-      new internal::mpi_cuda::GatherAlState<T>(
-        sendbuf, recvbuf, count, root, comm, stream);
-    internal::get_progress_engine()->enqueue(state);
-  }
-
-  template <typename T>
-  static void do_reduce(
-    const T* sendbuf, T* recvbuf, size_t count, ReductionOperator op,
-    int root, comm_type& comm,
-    cudaStream_t stream) {
-    internal::mpi_cuda::ReduceAlState<T>* state =
-      new internal::mpi_cuda::ReduceAlState<T>(
-        sendbuf, recvbuf, count, op, root, comm, stream);
-    internal::get_progress_engine()->enqueue(state);
-  }
-
-  template <typename T>
-  static void do_reduce_scatter(
-    const T* sendbuf, T* recvbuf, size_t count, ReductionOperator op,
-    comm_type& comm, cudaStream_t stream) {
-    internal::mpi_cuda::ReduceScatterAlState<T>* state =
-      new internal::mpi_cuda::ReduceScatterAlState<T>(
-        sendbuf, recvbuf, count, op, comm, stream);
-    internal::get_progress_engine()->enqueue(state);
-  }
-
-  template <typename T>
-  static void do_scatter(
-    const T* sendbuf, T* recvbuf, size_t count, int root, comm_type& comm,
-    cudaStream_t stream) {
-    internal::mpi_cuda::ScatterAlState<T>* state =
-      new internal::mpi_cuda::ScatterAlState<T>(
-        sendbuf, recvbuf, count, root, comm, stream);
-    internal::get_progress_engine()->enqueue(state);
-  }
 };
 
 template <>
