+++ test/test_transfer_to_one.cpp
@@ -4,8 +4,8 @@
 #ifdef AL_HAS_NCCL
 #include "test_utils_nccl_cuda.hpp"
 #endif
-#ifdef AL_HAS_MPI_CUDA
-#include "test_utils_mpi_cuda.hpp"
+#ifdef AL_HAS_HOST_TRANSFER
+#include "test_utils_ht.hpp"
 #endif
 
 #include <stdlib.h>
@@ -132,7 +132,12 @@ int main(int argc, char** argv) {
 #endif
 #ifdef AL_HAS_MPI_CUDA
   } else if (backend == "MPI-CUDA") {
-    test_correctness<Al::MPICUDABackend>();
+    std::cerr << "Point-to-point not supported on MPI-CUDA backend." << std::endl;
+    std::abort();
+#endif
+#ifdef AL_HAS_MPI_CUDA
+  } else if (backend == "HT") {
+    test_correctness<Al::HostTransferBackend>();
 #endif
   }
